1204_mops/requirements.txt
pymongo==4.6.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.3.0
aiohttp==3.10.11
python-dotenv==1.0.0
selenium==4.27.1
webdriver-manager==4.0.2
playwright==1.49.1
FinMind==1.5.1
